+++ packages/moonstone/Panels/CancelDecorator.js
@@ -1,5 +1,6 @@
 import hoc from '@enact/core/hoc';
 import Cancelable from '@enact/ui/Cancelable';
+import Spotlight from '@enact/spotlight';
 
 const defaultConfig = {
 	cancel: null
@@ -11,6 +12,12 @@ const CancelDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	function handleCancel (props) {
 		const {index, [cancel]: handler} = props;
 		if (index > 0 && handler) {
+			// clear Spotlight focus
+			const current = Spotlight.getCurrent();
+			if (current) {
+				current.blur();
+			}
+
 			handler({
 				index: index - 1
 			});