+++ packages/moonstone/Scroller/ScrollableNative.js
@@ -271,7 +271,11 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		componentDidUpdate () {
 			// Need to sync calculated client size if it is different from the real size
 			if (this.childRef.syncClientSize) {
-				this.childRef.syncClientSize();
+				// If we actually synced, we need to reset scroll position.
+				if (this.childRef.syncClientSize()) {
+					this.setScrollLeft(0);
+					this.setScrollTop(0);
+				}
 			}
 
 			this.direction = this.childRef.props.direction;
@@ -307,7 +311,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				// FIXME `onScroll` doesn't work on the v8 snapshot.
 				childContainerRef.removeEventListener('scroll', this.onScroll, {capture: true});
 				// FIXME `onFocus` doesn't work on the v8 snapshot.
-				childContainerRef.removeEventListener('focus', this.onFocus, {capture: true});
+				childContainerRef.removeEventListener('focusin', this.onFocus);
 				// FIXME `onMouseOver` doesn't work on the v8 snapshot.
 				childContainerRef.removeEventListener('mouseover', this.onMouseOver, {capture: true});
 				// FIXME `onMouseMove` doesn't work on the v8 snapshot.
@@ -322,7 +326,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		isScrollAnimationTargetAccumulated = false
 		deferScrollTo = true
 		pageDistance = 0
-		animateOnFocus = true
+		animateOnFocus = false
 		pageDirection = 0
 
 		// event handlers
@@ -544,20 +548,30 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		}
 
 		scrollByPage = (keyCode) => {
+			// Only scroll by page when the vertical scrollbar is visible. Otherwise, treat the
+			// scroller as a plain container
+			if (!this.state.isVerticalScrollbarVisible) return;
+
 			const
 				{getEndPoint, scrollToAccumulatedTarget} = this,
 				bounds = this.getScrollBounds(),
 				canScrollVertically = this.canScrollVertically(bounds),
+				childRef = this.childRef,
 				pageDistance = isPageUp(keyCode) ? (this.pageDistance * -1) : this.pageDistance,
 				spotItem = Spotlight.getCurrent();
 
 			if (!Spotlight.getPointerMode() && spotItem) {
 				// Should skip scroll by page when spotItem is paging control button of Scrollbar
-				if (!this.childRef.containerRef.contains(spotItem)) {
+				if (!childRef.containerRef.contains(spotItem)) {
 					return;
 				}
 				const
-					containerId = Spotlight.getActiveContainer(),
+					containerId = (
+						// ScrollerNative has a containerId on containerRef
+						childRef.containerRef.dataset.containerId ||
+						// VirtualListNative has a containerId on contentRef
+						childRef.contentRef.dataset.containerId
+					),
 					direction = this.getPageDirection(keyCode),
 					rDirection = reverseDirections[direction],
 					viewportBounds = this.containerRef.getBoundingClientRect(),
@@ -565,17 +579,22 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					endPoint = getEndPoint(direction, spotItemBounds, viewportBounds),
 					next = getTargetByDirectionFromPosition(rDirection, endPoint, containerId);
 
+				// If there is no next spottable DOM elements, scroll one page with animation
 				if (!next) {
 					scrollToAccumulatedTarget(pageDistance, canScrollVertically);
+				// If there is a next spottable DOM element vertically or horizontally, focus it without animation
 				} else if (next !== spotItem) {
 					this.animateOnFocus = false;
 					Spotlight.focus(next);
+				// If a next spottable DOM element is equals to the current spottable item, we need to find a next item
 				} else {
-					const nextPage = this.childRef.scrollToNextPage({direction, reverseDirection: rDirection, focusedItem: spotItem});
+					const nextPage = childRef.scrollToNextPage({direction, reverseDirection: rDirection, focusedItem: spotItem, containerId});
 
+					// If finding a next spottable item in a Scroller, focus it
 					if (typeof nextPage === 'object') {
 						this.animateOnFocus = false;
 						Spotlight.focus(nextPage);
+					// Scroll one page with animation if nextPage is equals to `false`
 					} else if (!nextPage) {
 						scrollToAccumulatedTarget(pageDistance, canScrollVertically);
 					}
@@ -795,6 +814,11 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				} else {
 					if (typeof opt.index === 'number' && typeof this.childRef.getItemPosition === 'function') {
 						itemPos = this.childRef.getItemPosition(opt.index, opt.stickTo);
+						// If the first or the last item to the sticked direction is disabled,
+						// focus another item which is enabled.
+						if (opt.nodeIndexToBeFocused) {
+							this.childRef.setNodeIndexToBeFocused(opt.nodeIndexToBeFocused);
+						}
 					} else if (opt.node instanceof Object) {
 						if (opt.node.nodeType === 1 && typeof this.childRef.getNodePosition === 'function') {
 							itemPos = this.childRef.getNodePosition(opt.node);
@@ -967,7 +991,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				// FIXME `onScroll` doesn't work on the v8 snapshot.
 				childContainerRef.addEventListener('scroll', this.onScroll, {capture: true});
 				// FIXME `onFocus` doesn't work on the v8 snapshot.
-				childContainerRef.addEventListener('focus', this.onFocus, {capture: true});
+				childContainerRef.addEventListener('focusin', this.onFocus);
 				// FIXME `onMouseOver` doesn't work on the v8 snapshot.
 				childContainerRef.addEventListener('mouseover', this.onMouseOver, {capture: true});
 				// FIXME `onMouseMove` doesn't work on the v8 snapshot.