+++ packages/moonstone/Scroller/Scroller.js
@@ -123,6 +123,8 @@ class ScrollerBase extends Component {
 		left: 0
 	}
 
+	isScrolledToBoundary = false
+
 	getScrollBounds = () => this.scrollBounds
 
 	getRtlPositionX = (x) => (this.context.rtl ? this.scrollBounds.maxLeft - x : x)
@@ -199,12 +201,16 @@ class ScrollerBase extends Component {
 		return node.getBoundingClientRect();
 	}
 
-	scrollToNextPage = ({direction, reverseDirection, focusedItem}) => {
+	scrollToNextPage = ({direction, reverseDirection, focusedItem, containerId}) => {
 		const
 			endPoint = this.getNextEndPoint(direction, focusedItem.getBoundingClientRect()),
-			next = getTargetByDirectionFromPosition(reverseDirection, endPoint, Spotlight.getActiveContainer());
+			next = getTargetByDirectionFromPosition(reverseDirection, endPoint, containerId);
 
-		return (next !== focusedItem) && next;
+		if (next === focusedItem) {
+			return false; // Scroll one page with animation
+		} else {
+			return next; // Focus a next item
+		}
 	}
 
 	getNextEndPoint = (direction, oSpotBounds) => {
@@ -257,9 +263,7 @@ class ScrollerBase extends Component {
 
 		if (this.isVertical()) {
 			this.scrollPos.top = this.calculateScrollTop(item, itemTop, itemHeight, scrollInfo, scrollPosition);
-		}
-
-		if (this.isHorizontal()) {
+		} else if (this.isHorizontal()) {
 			const
 				{clientWidth} = this.scrollBounds,
 				rtlDirection = this.context.rtl ? -1 : 1,
@@ -423,9 +427,13 @@ class ScrollerBase extends Component {
 
 		const {keyCode, target} = ev;
 		const direction = getDirection(keyCode);
+		if (!ev.repeat) {
+			this.isScrolledToBoundary = false;
+		}
 
-		if (direction && !this.findInternalTarget(direction, target)) {
+		if (direction && !this.findInternalTarget(direction, target) && !this.isScrolledToBoundary) {
 			this.scrollToBoundary(direction);
+			this.isScrolledToBoundary = true;
 		}
 	}