+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -11,6 +11,7 @@ import equals from 'ramda/src/equals';
 import React from 'react';
 import PropTypes from 'prop-types';
 import DurationFmt from '@enact/i18n/ilib/lib/DurationFmt';
+import {contextTypes, FloatingLayerDecorator} from '@enact/ui/FloatingLayer';
 import {forKey, forward, forwardWithPrevent, handle, stopImmediate} from '@enact/core/handle';
 import ilib from '@enact/i18n';
 import {Job} from '@enact/core/util';
@@ -40,7 +41,10 @@ import Times from './Times';
 import css from './VideoPlayer.less';
 
 const SpottableDiv = Touchable(Spottable('div'));
-const Container = SpotlightContainerDecorator({enterTo: ''}, 'div');
+const Container = SpotlightContainerDecorator(
+	{enterTo: ''},
+	'div'
+);
 
 // Keycode map for webOS TV
 const keyMap = {
@@ -168,6 +172,8 @@ const AnnounceState = {
  */
 const VideoPlayerBase = class extends React.Component {
 	static displayName = 'VideoPlayerBase'
+
+	static contextTypes = contextTypes
 
 	static propTypes = /** @lends moonstone/VideoPlayer.VideoPlayerBase.prototype */ {
 		/**
@@ -455,6 +461,13 @@ const VideoPlayerBase = class extends React.Component {
 		onScrub: PropTypes.func,
 
 		/**
+		 * Function executed when seek is attemped while `seekDisabled` is true.
+		 *
+		 * @type {Function}
+		 */
+		onSeekFailed: PropTypes.func,
+
+		/**
 		 * When `true`, the video will pause when it reaches either the start or the end of the
 		 * video during rewind, slow rewind, fast forward, or slow forward.
 		 *
@@ -518,6 +531,16 @@ const VideoPlayerBase = class extends React.Component {
 		 * @public
 		 */
 		rightComponents: PropTypes.node,
+
+		/**
+		 * When `true`, seek function is disabled.
+		 *
+		 * Note that jump by arrow keys will also be disabled when `true`.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		seekDisabled: PropTypes.bool,
 
 		/**
 		 * Registers the VideoPlayer component with an
@@ -737,6 +760,8 @@ const VideoPlayerBase = class extends React.Component {
 			this.reloadVideo();
 		}
 
+		this.setFloatingLayerShowing(this.state.mediaControlsVisible || this.state.mediaSliderVisible);
+
 		// Added to set default focus on the media control (play) when controls become visible.
 		if (
 			this.state.mediaControlsVisible &&
@@ -762,6 +787,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.renderBottomControl.stop();
 		this.stopListeningForPulses();
 		this.sliderTooltipTimeJob.stop();
+		this.slider5WayPressJob.stop();
 	}
 
 	//
@@ -856,6 +882,13 @@ const VideoPlayerBase = class extends React.Component {
 		}
 
 		return true;
+	}
+
+	setFloatingLayerShowing = (showing) => {
+		const layer = this.context.getFloatingLayer && this.context.getFloatingLayer();
+		if (layer) {
+			layer.style.display = showing ? 'block' : 'none';
+		}
 	}
 
 	/**
@@ -1033,7 +1066,9 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	doPulseAction () {
-		if (is('left', this.pulsingKeyCode)) {
+		if (this.props.seekDisabled) {
+			forward('onSeekFailed', {}, this.props);
+		} else if (is('left', this.pulsingKeyCode)) {
 			this.showMiniFeedback = true;
 			this.jump(-1 * this.props.jumpBy);
 			this.announceJob.startAfter(500, secondsToTime(this.video.currentTime, this.durfmt, {includeHour: true}));
@@ -1252,7 +1287,11 @@ const VideoPlayerBase = class extends React.Component {
 	 * @public
 	 */
 	seek = (timeIndex) => {
-		this.video.currentTime = timeIndex;
+		if (!this.props.seekDisabled) {
+			this.video.currentTime = timeIndex;
+		} else {
+			forward('onSeekFailed', {}, this.props);
+		}
 	}
 
 	/**
@@ -1620,6 +1659,18 @@ const VideoPlayerBase = class extends React.Component {
 		});
 	}
 
+	slider5WayPressJob = new Job(() => {
+		this.setState({slider5WayPressed: false});
+	}, 200);
+
+	handleSliderKeyDown = (ev) => {
+		if (is('enter', ev.keyCode)) {
+			this.setState({
+				slider5WayPressed: true
+			}, this.slider5WayPressJob.start());
+		}
+	}
+
 	onJumpBackward = this.handle(
 		(ev, props) => forwardJumpBackwardButtonClick(this.addStateToEvent(ev), props),
 		() => this.jump(-1 * this.props.jumpBy)
@@ -1746,15 +1797,17 @@ const VideoPlayerBase = class extends React.Component {
 		delete rest.jumpBy;
 		delete rest.jumpDelay;
 		delete rest.no5WayJump;
-		delete rest.onControlsAvailable;
 		delete rest.onBackwardButtonClick;
+		delete rest.onControlsAvailable;
 		delete rest.onForwardButtonClick;
 		delete rest.onJumpBackwardButtonClick;
 		delete rest.onJumpForwardButtonClick;
 		delete rest.onPlayButtonClick;
 		delete rest.onScrub;
+		delete rest.onSeekFailed;
 		delete rest.pauseAtEnd;
 		delete rest.playbackRateHash;
+		delete rest.seekDisabled;
 		delete rest.setApiProvider;
 		delete rest.thumbnailUnavailable;
 		delete rest.titleHideDelay;
@@ -1770,7 +1823,7 @@ const VideoPlayerBase = class extends React.Component {
 		const controlsAriaProps = this.getControlsAriaProps();
 
 		return (
-			<div className={css.videoPlayer + (className ? ' ' + className : '')} style={style} onClick={this.activityDetected} onKeyDown={this.activityDetected} ref={this.setPlayerRef}>
+			<div className={css.videoPlayer + ' enact-fit' + (className ? ' ' + className : '')} style={style} onClick={this.activityDetected} onKeyDown={this.activityDetected} ref={this.setPlayerRef}>
 				{/* Video Section */}
 				<video
 					{...rest}
@@ -1827,14 +1880,16 @@ const VideoPlayerBase = class extends React.Component {
 
 							{noSlider ? null : <MediaSlider
 								backgroundProgress={this.state.proportionLoaded}
-								value={this.state.proportionPlayed}
 								onBlur={this.handleSliderBlur}
 								onChange={this.onSliderChange}
 								onFocus={this.handleSliderFocus}
+								onKeyDown={this.handleSliderKeyDown}
 								onKnobMove={this.handleKnobMove}
-								onSpotlightUp={this.handleSpotlightUpFromSlider}
 								onSpotlightDown={this.handleSpotlightDownFromSlider}
+								onSpotlightUp={this.handleSpotlightUpFromSlider}
+								forcePressed={this.state.slider5WayPressed}
 								spotlightDisabled={spotlightDisabled || this.state.mediaSliderVisible && !this.state.mediaControlsVisible}
+								value={this.state.proportionPlayed}
 								visible={this.state.mediaSliderVisible}
 							>
 								<FeedbackTooltip
@@ -1953,8 +2008,11 @@ const VideoPlayer = ApiDecorator(
 	{api: ['fastForward', 'getMediaState', 'hideControls', 'jump', 'pause', 'play', 'rewind', 'seek', 'showControls']},
 	Slottable(
 		{slots: ['infoComponents', 'leftComponents', 'rightComponents', 'source']},
-		Skinnable(
-			VideoPlayerBase
+		FloatingLayerDecorator(
+			{floatLayerId:  'videoPlayerFloatingLayer'},
+			Skinnable(
+				VideoPlayerBase
+			)
 		)
 	)
 );