+++ packages/moonstone/internal/Picker/Picker.js
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types';
 import Touchable from '@enact/ui/Touchable';
 import shouldUpdate from 'recompose/shouldUpdate';
 import {SlideLeftArranger, SlideTopArranger, ViewManager} from '@enact/ui/ViewManager';
-import {getDirection} from '@enact/spotlight';
+import Spotlight, {getDirection} from '@enact/spotlight';
 
 import Skinnable from '../../Skinnable';
 import {validateRange, validateStepped} from '../validators';
@@ -374,10 +374,10 @@ const PickerBase = class extends React.Component {
 		}
 	}
 
-	componentDidUpdate () {
-		if (this.props.joined) {
+	componentDidUpdate (prevProps) {
+		if (this.props.joined === true && prevProps.joined === false) {
 			this.containerRef.addEventListener('wheel', this.handleWheel);
-		} else {
+		} else if (prevProps.joined === true && this.props.joined === false) {
 			this.containerRef.removeEventListener('wheel', this.handleWheel);
 		}
 	}
@@ -481,10 +481,12 @@ const PickerBase = class extends React.Component {
 	}
 
 	handleWheel = (ev) => {
-		const {joined, step} = this.props;
+		const {step} = this.props;
 		forwardWheel(ev, this.props);
 
-		if (joined) {
+		const isContainerSpotted = this.containerRef === Spotlight.getCurrent();
+
+		if (isContainerSpotted) {
 			const dir = -Math.sign(ev.deltaY);
 
 			// We'll sometimes get a 0/-0 wheel event we need to ignore or the wheel event has reached
@@ -504,20 +506,6 @@ const PickerBase = class extends React.Component {
 				ev.preventDefault();
 				ev.stopPropagation();
 			}
-		}
-	}
-
-	handleDecPulse = () => {
-		if (!this.hasReachedBound(this.props.step * -1)) {
-			this.handleDown(-1);
-			this.updateValue(-1);
-		}
-	}
-
-	handleIncPulse = () => {
-		if (!this.hasReachedBound(this.props.step)) {
-			this.handleDown(1);
-			this.updateValue(1);
 		}
 	}
 
@@ -758,7 +746,7 @@ const PickerBase = class extends React.Component {
 			sizingPlaceholder = <div aria-hidden className={css.sizingPlaceholder}>{ '0'.repeat(width) }</div>;
 		}
 
-		const valueText = this.calcValueText();
+		const valueText = ariaValueText != null ? ariaValueText : this.calcValueText();
 		const decrementerAriaControls = !incrementerDisabled ? id : null;
 		const incrementerAriaControls = !decrementerDisabled ? id : null;
 
@@ -779,14 +767,14 @@ const PickerBase = class extends React.Component {
 			>
 				<PickerButton
 					aria-controls={!joined ? incrementerAriaControls : null}
-					aria-label={this.calcIncrementLabel(ariaValueText != null ? ariaValueText : valueText)}
+					aria-label={this.calcIncrementLabel(valueText)}
 					className={css.incrementer}
 					disabled={incrementerDisabled}
 					hidden={reachedEnd}
 					icon={incrementIcon}
 					joined={joined}
 					onDown={this.handleIncDown}
-					onHoldPulse={this.handleIncPulse}
+					onHoldPulse={this.handleIncDown}
 					onKeyDown={this.handleIncKeyDown}
 					onSpotlightDisappear={onIncrementSpotlightDisappear}
 					onTap={this.handleIncClick}
@@ -795,7 +783,7 @@ const PickerBase = class extends React.Component {
 				<div
 					aria-disabled={disabled}
 					aria-hidden={!active}
-					aria-valuetext={ariaValueText != null ? ariaValueText : valueText}
+					aria-valuetext={valueText}
 					className={css.valueWrapper}
 					id={id}
 					role="spinbutton"
@@ -814,14 +802,14 @@ const PickerBase = class extends React.Component {
 				</div>
 				<PickerButton
 					aria-controls={!joined ? decrementerAriaControls : null}
-					aria-label={this.calcDecrementLabel(ariaValueText != null ? ariaValueText : valueText)}
+					aria-label={this.calcDecrementLabel(valueText)}
 					className={css.decrementer}
 					disabled={decrementerDisabled}
 					hidden={reachedStart}
 					icon={decrementIcon}
 					joined={joined}
 					onDown={this.handleDecDown}
-					onHoldPulse={this.handleDecPulse}
+					onHoldPulse={this.handleDecDown}
 					onKeyDown={this.handleDecKeyDown}
 					onSpotlightDisappear={onDecrementSpotlightDisappear}
 					onTap={this.handleDecClick}