+++ packages/core/CHANGELOG.md
@@ -20,6 +20,24 @@ No significant changes.
 
 No significant changes.
 
+## [1.12.1] - 2017-11-07
+
+No significant changes.
+
+## [1.12.0] - 2017-10-27
+
+### Added
+
+- `core/util.Job` APIs `idle`, `idleUntil`, `startRaf` and `startRafAfter`
+
+## [1.11.0] - 2017-10-24
+
+No significant changes.
+
+## [1.10.1] - 2017-10-16
+
+No significant changes.
+
 ## [1.10.0] - 2017-10-09
 
 No significant changes.
@@ -130,7 +148,7 @@ No significant changes.
 
 ## [1.0.0-beta.3] - 2017-02-21
 
-### Addded
+### Added
 
 - `core/handle` function `forEventProp` to test properties on an event