+++ packages/moonstone/CHANGELOG.md
@@ -12,6 +12,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
+- `moonstone/VirtualList` to sync scroll position when clientSize changed
+
 ## [2.0.0-alpha.2] - 2017-08-29
 
 No significant changes.
@@ -21,6 +23,67 @@ No significant changes.
 ### Changed
 
 - `moonstone/Button`, `moonstone/Checkbox`, `moonstone/FormCheckbox`, `moonstone/IconButton`, `moonstone/IncrementSlider`, `moonstone/Item`, `moonstone/Picker`, and `moonstone/RangePicker`, `moonstone/Switch` and `moonstone/VideoPlayer` to use `ui/Touchable`
+
+## [1.12.1] - 2017-11-07
+
+### Fixed
+
+- `moonstone/ExpandableItem` and all other `Expandable` components to now pull down from the top instead of being revealed from the bottom, matching Enyo's design
+- `moonstone/VirtualListNative` to scroll properly with page up/down keys if there is an disabled item
+- `moonstone/RangePicker` to display negative values correctly in RTL
+- `moonstone/Scrollable` to not blur scroll buttons when wheeling
+- `moonstone/Scrollbar` to hide scroll thumb immediately without delay after scroll position to be min or max
+- `moonstone/Divider` to pass `marqueeOn` prop
+- `moonstone/Slider` to fire `onChange` on mouse up and key up
+- `moonstone/VideoPlayer` to show knob when pressed
+- `moonstone/Header` to layout `titleBelow` and `subTitleBelow` correctly
+- `moonstone/Header` to use correct font-weight for `subTitleBelow`
+
+## [1.12.0] - 2017-10-27
+
+### Fixed
+
+- `moonstone/Scrollable` to prevent focusing outside the viewport when pressing a 5-way key during wheeling
+- `moonstone/Scroller` to called scrollToBoundary once when focus is moved using holding child item
+- `moonstone/VideoPlayer` to apply skin correctly
+- `moonstone/Popup` from `last-focused` to `default-element` in `SpotlightContainerDecorator` config
+- `moonstone/Panels` to retain focus when back key is pressed on breadcrumb
+- `moonstone/Input` to correctly hide VKB when dismissing
+
+## [1.11.0] - 2017-10-24
+
+### Added
+
+- `moonstone/VideoPlayer` properties `seekDisabled` and `onSeekFailed` to disable seek function
+
+### Changed
+
+- `moonstone/ExpandableList` to become `disabled` if there are no children
+
+### Fixed
+
+- `moonstone/Picker` to read out customized accessibility value when picker prop has `joined` and `aria-valuetext`
+- `moonstone/Scroller` to apply scroll position on vertical or horizontal Scroller when child gets a focus
+- `moonstone/Scroller.Scrollable` to scroll withtout animation when panel is changed
+- `moonstone/ContextualPopup` padding to not overlap close button
+- `moonstone/Scroller.Scrollable` and `moonstone/Scroller` to change focus via page up/down only when the scrollbar is visible
+- `moonstone/Picker` to only increment one value on hold
+- `moonstone/ItemOverlay` to remeasure when focused
+
+## [1.10.1] - 2017-10-16
+
+### Fixed
+
+- `moonstone/Scrollable` and `moonstone/Scroller` to scroll via page up/down when focus is inside a Spotlight container
+- `moonstone/VirtualList` and `moonstone/VirtualGridList` to scroll by 5-way keys right after wheeling
+- `moonstone/VirtualList` not to move focus when a current item and the last item are located at the same line and pressing a page down key
+- `moonstone/Slider` knob to follow while dragging for detached knob
+- `moonstone/Header` to layout header row correctly in `standard` type
+- `moonstone/Input` to not dismiss on-screen keyboard when dragging cursor out of input box
+- `moonstone/Header` RTL `line-height` issue
+- `moonstone/Panels` to render children on idle
+- `moonstone/Scroller.Scrollable` to limit its muted spotlight container scrim to its bounds
+- `moonstone/Input` to always forward `onKeyUp` event
 
 ## [1.10.0] - 2017-10-09
 
@@ -42,10 +105,13 @@ No significant changes.
 - `moonstone/Slider` toggling activated state when holding enter/select key
 - `moonstone/TimePicker` picker icons shifting slightly when focusing an adjacent picker
 - `moonstone/Icon` so it handles color the same way generic text does, by inheriting from the parent's color. This applies to all instances of `Icon`, `IconButton`, and `Icon` inside `Button`.
+- `moonstone/fonts` Museo Sans font to correct "Ti" kerning
 - `moonstone/VideoPlayer` to correctly position knob on mouse click
 - `moonstone/Panels.Header` to show an ellipsis for long titles with RTL text
 - `moonstone/Marquee` to restart when invalidated by a prop change and managed by a `moonstone/Marquee.MarqueeController`
 - `spotlight.Spotlight` method `focus()` to verify that the target element matches its container's selector rules prior to setting focus
+- `moonstone/Picker` to only change picker values `onWheel` when spotted
+- `moonstone/VideoPlayer` to hide descendant floating components (tooltips, contextual popups) when the media controls hide
 
 ## [1.9.3] - 2017-10-03