+++ packages/sampler/CHANGELOG.md
@@ -20,6 +20,22 @@ No significant changes.
 
 No significant changes.
 
+## [1.12.1] - 2017-11-07
+
+No significant changes.
+
+## [1.12.0] - 2017-10-27
+
+No significant changes.
+
+## [1.11.0] - 2017-10-24
+
+No significant changes.
+
+## [1.10.1] - 2017-10-16
+
+No significant changes.
+
 ## [1.10.0] - 2017-10-09
 
 No significant changes.